requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
networkx>=3.1
matplotlib>=3.8
pyyaml>=6.0
xgboost>=2.0
torch>=2.0
tqdm>=4.0
